pyproject.toml
[tool.poetry]
name = "idoven-challenge"
version = "0.1.0"
description = ""
authors = ["Manu Molina <manu.molinam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.95.1"
requests = "^2.28.1"
uvicorn = { extras = ["standard"], version = "^0.20.0" }
gunicorn = "^21.2.0"
psycopg2-binary = "2.9.6"
sqlmodel = "0.0.8"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
PyJWT = "~2.8.0"
passlib = "~1.7.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.2"
coverage = "^7.2.6"
pytest-coverage = "^0.0"
pytest-mock = "^3.10.0"
ruff = { extras = ["toml"], version = "==0.0.265" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"